requirements.txt
anyio==3.6.2
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2
slack-bolt==1.16.1
slack-sdk==3.19.5
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
